AI-Bookmark/server/requirements.txt
Flask==2.2.2
flask-cors
neo4j==5.7.0
requests
beautifulsoup4
python-dotenv
google-generativeai